An fucked attempt to draw a cube with many small triangles
So that the cube itself would be warped by the barrel shader. The problem I
have with this commit, is I can not get it to render anything more than the
first face. If I comment out the first front and back face, the code *does*
render the left face! So basically, whatever I do I can only get it to render
the first face that is loaded (which ever that is), proving the geometry is
ok.

My current theory is that triangles in a glDrawElements call must be
"touching" each other. So when I render the front face, then the back face,
those two faces are separated by empty space, therefore for some reason the
element renderer stops rendering when it hits a "break" in the geometry.

So I am giving up at this point because I am going to instead render to a
texture, render the texture onto a grid of triangles, and warp that grid.
This is mainly so I can render primitive shapes, without having to break
them up into hundreds of triangles so they warp properly.